+++ src/engine/network/utils/thread.ts
@@ -4,7 +4,7 @@ import {writable} from "@coracle.social/lib"
 import {Tags, getIdOrAddress, getIdFilters, getIdAndAddress} from "@coracle.social/util"
 import type {DisplayEvent} from "src/engine/notes/model"
 import type {Event} from "src/engine/events/model"
-import {useRelaysWithFallbacks} from "src/engine/relays/utils"
+import {withFallbacks} from "src/engine/relays/utils"
 import {load} from "./executor"
 
 const getAncestorIds = e =>
@@ -39,7 +39,7 @@ export class ThreadLoader {
 
     if (filteredIds.length > 0) {
       load({
-        relays: useRelaysWithFallbacks(this.relays),
+        relays: withFallbacks(this.relays),
         filters: getIdFilters(filteredIds),
         onEvent: batch(300, (events: Event[]) => {
           this.addToThread(events)