+++ src/engine/relays/commands.ts
@@ -4,10 +4,11 @@ import {Tag, normalizeRelayUrl, createEvent, isShareableRelayUrl} from "@coracle
 import {people} from "src/engine/people/state"
 import {session, canSign, signer, stateKey} from "src/engine/session/derived"
 import {updateStore} from "src/engine/core/commands"
-import {createAndPublish, getClientTags, Publisher} from "src/engine/network/utils"
+import {getClientTags, Publisher} from "src/engine/network/utils"
 import type {RelayPolicy} from "./model"
 import {relays} from "./state"
 import {relayPolicies} from "./derived"
+import {hints, forcePlatformRelays, withIndexers} from "./utils"
 
 export const saveRelay = (url: string) => {
   url = normalizeRelayUrl(url)
@@ -38,41 +39,47 @@ export const saveRelayPolicy = (e, relays: RelayPolicy[]) => {
 }
 
 export const publishRelays = ($relays: RelayPolicy[]) => {
-  console.log($relays)
+  updateStore(people.key(stateKey.get()), now(), {relays: $relays})
+
   if (canSign.get()) {
-    return createAndPublish(10002, {
-      tags: $relays
-        .filter(r => isShareableRelayUrl(r.url))
-        .flatMap(r => {
-          const tag = Tag.from(["r", normalizeRelayUrl(r.url)])
+    const relays = withIndexers(forcePlatformRelays(hints.WriteRelays().getUrls()))
+    const event = signer.get().signAsUser(
+      createEvent(10002, {
+        tags: $relays
+          .filter(r => isShareableRelayUrl(r.url))
+          .flatMap(r => {
+            const tag = Tag.from(["r", normalizeRelayUrl(r.url)])
 
-          if (r.read && r.write) return [tag.valueOf()]
-          if (r.write) return [tag.append("write").valueOf()]
-          if (r.read) return [tag.append("read").valueOf()]
+            if (r.read && r.write) return [tag.valueOf()]
+            if (r.write) return [tag.append("write").valueOf()]
+            if (r.read) return [tag.append("read").valueOf()]
 
-          return []
-        })
-        .concat(getClientTags()),
-    })
-  } else {
-    updateStore(people.key(stateKey.get()), now(), {relays: $relays})
+            return []
+          })
+          .concat(getClientTags()),
+      }),
+    )
+
+    Publisher.publish({event, relays})
   }
+}
+
+export const requestRelayAccess = async (url: string, claim: string, sk?: string) => {
+  const template = createEvent(28934, {tags: [["claim", claim]]})
+  const event = sk
+    ? await signer.get().signWithKey(template, sk)
+    : await signer.get().signAsUser(template)
+
+  const pub = Publisher.publish({relays: [url], event})
+
+  await pub.result
 }
 
 export const joinRelay = async (url: string, claim?: string) => {
   url = normalizeRelayUrl(url)
 
   if (claim && canSign.get()) {
-    const pub = Publisher.publish({
-      relays: [url],
-      event: await signer.get().signAsUser(
-        createEvent(28934, {
-          tags: [["claim", claim]],
-        }),
-      ),
-    })
-
-    await pub.result
+    await requestRelayAccess(url, claim)
   }
 
   // Re-publish user meta to the new relay